Music.Dynamics.Literal

Description

Provides overloaded pitch literals.

Documentation

newtype DynamicsL

Dynamics literal.

First value is start value, second is end value.

(Just x, Nothing) is a constant dynamic of x * (Just x, Just y) is a dynamic varying from x to y * (Nothing, Just y) is a dynamic varying from the previous level to y * (Nothing, Nothing) is a dynamic varying from the previous level to the next.

For levels, we use -0.5 for mp, 0.5 for mf and add or remove one for each level. 0 is an unspecified middle level dynamic.

Constructors

DynamicsL

Fields

getDynamicsL :: (Maybe Double, Maybe Double)

class IsDynamics a where

Methods

fromDynamics :: DynamicsL -> a

pppppp :: IsDynamics a => a

ppppp :: IsDynamics a => a

pppp :: IsDynamics a => a

ppp :: IsDynamics a => a

pp :: IsDynamics a => a

_p :: IsDynamics a => a

mp :: IsDynamics a => a

mf :: IsDynamics a => a

_f :: IsDynamics a => a

ff :: IsDynamics a => a

fff :: IsDynamics a => a

ffff :: IsDynamics a => a

fffff :: IsDynamics a => a

ffffff :: IsDynamics a => a

sffz :: IsDynamics a => a

sfz :: IsDynamics a => a

fz :: IsDynamics a => a

rfz :: IsDynamics a => a

fp :: IsDynamics a => a