Music.MusicXml.Simple

Description

Provides smart constructors for the MusicXML representation.

Documentation

module Music.MusicXml

Score and parts

Basic constructors

fromPart :: String -> String -> String -> [Music] -> Score

Create a single-part score.

 fromPart title composer partName measures

Example:

 fromPart "Suite" "Bach" "Cello solo" []

fromParts :: String -> String -> PartList -> [[Music]] -> Score

Create a multi-part score.

 fromParts title composer partList parts

Example:

 fromParts "4'33" "Cage" (partList ["Violin", "Viola", "Cello"]) [[]]

Part lists

partList :: [String] -> PartList

Create a part list from instrument names.

partListAbbr :: [(String, String)] -> PartList

Create a part list from instrument names and abbreviations.

bracket :: PartList -> PartList

Enclose the given parts in a bracket.

brace :: PartList -> PartList

Enclose the given parts in a brace.

Measures

measure :: [Music] -> Music

Convenient synonym for mconcat, allowing us to write things like

 measure [
    beam [ 
        note c  (1/8), 
        note d  (1/8),
        note e  (1/8),
        note f  (1/8) 
    ],
    tuplet 3 2 [ 
        note g  (1/4),
        note a  (1/4),
        note b  (1/4) 
    ]
 ]

bar :: [Music] -> Music

Convenient synonym for mconcat.

Top-level attributes

Pitch

trebleClef :: Music

altoClef :: Music

bassClef :: Music

defaultClef :: Music

clef :: ClefSign -> Line -> Music

Create a clef.

defaultKey :: Music

key :: Fifths -> Mode -> Music

Create a key signature.

Time

defaultDivisions :: Music

Set the tick division to the default value.

divisions :: Divs -> Music

Define the number of ticks per quarter note.

commonTime :: Music

cutTime :: Music

time :: Beat -> BeatType -> Music

Create a time signature.

Tempo

metronome :: NoteVal -> Tempo -> Music

Create a metronome mark.

metronome' :: NoteVal -> Bool -> Tempo -> Music

Create a metronome mark.

Backup and forward

backup :: Duration -> Music

forward :: Duration -> Music

Notes

Basic constructors

rest :: NoteVal -> Music

Create a rest.

 rest (1/4)
 rest (3/8)
 rest quarter
 rest (dotted eight)

note :: Pitch -> NoteVal -> Music

Create a single note.

 note c   (1/4)
 note fs_ (3/8)
 note c   quarter
 note (c + pure fifth) (dotted eight)

chord :: [Pitch] -> NoteVal -> Music

Create a chord.

 chord [c,eb,fs_] (3/8)
 chord [c,d,e] quarter
 chord [c,d,e] (dotted eight)

Voice

setVoice :: Int -> Music -> Music

Duration

dot :: Music -> Music

tuplet :: Int -> Int -> Music -> Music

setNoteVal :: NoteVal -> Music -> Music

separateDots :: NoteVal -> (NoteVal, Int)

Beams

beam :: Music -> Music

beginBeam :: Music -> Music

continueBeam :: Music -> Music

endBeam :: Music -> Music

Ties

beginTie :: Music -> Music

endTie :: Music -> Music

Note heads

setNoteHead :: NoteHead -> Music -> Music

Notations

addNotation :: Notation -> Music -> Music

Pitch transformations

Glissando

beginGliss :: Music -> Music

endGliss :: Music -> Music

Slides

beginSlide :: Music -> Music

endSlide :: Music -> Music

Time transformations

Accelerando and ritardando

Fermatas and breaks

fermata :: FermataSign -> Music -> Music

breathMark :: Music -> Music

caesura :: Music -> Music

Articulation

Slurs

slur :: Music -> Music

beginSlur :: Music -> Music

endSlur :: Music -> Music

Staccato and tenuto

staccato :: Music -> Music

tenuto :: Music -> Music

spiccato :: Music -> Music

staccatissimo :: Music -> Music

Accents

accent :: Music -> Music

strongAccent :: Music -> Music

Miscellaneous

scoop :: Music -> Music

plop :: Music -> Music

doit :: Music -> Music

falloff :: Music -> Music

stress :: Music -> Music

unstress :: Music -> Music

Ornaments

trill :: Music -> Music

turn :: Bool -> Bool -> Music -> Music

shake :: Music -> Music

mordent :: Bool -> Music -> Music

tremolo :: Int -> Music -> Music

Dynamics

Crescendo and diminuendo

cresc :: Music -> Music

dim :: Music -> Music

beginCresc :: Music

endCresc :: Music

beginDim :: Music

endDim :: Music

Dynamic levels

dynamic :: Dynamics -> Music

Both

crescFrom :: Dynamics -> Music -> Music

crescTo :: Dynamics -> Music -> Music

crescFromTo :: Dynamics -> Dynamics -> Music -> Music

dimFrom :: Dynamics -> Music -> Music

dimTo :: Dynamics -> Music -> Music

dimFromTo :: Dynamics -> Dynamics -> Music -> Music

Text

text :: String -> Music

rehearsal :: String -> Music

segno :: Music

coda :: Music

Folds and maps

mapMusic :: (Attributes -> Attributes) -> (Note -> Note) -> (Direction -> Direction) -> Music -> Music

foldMusic :: Monoid m => (Attributes -> r) -> (Note -> r) -> (Direction -> r) -> (r -> m) -> Music -> m