Music.Parts

Documentation

data Division

A division represents a subset of a finite group of performers.

For example a group may be divided into three equal divisions, designated (0, 3), (1, 3) and (2, 3) respectively.

type Subpart = [Division]

A subpart is a potentially infinite sequence of divisions, each typically designated using a new index type, i.e. I.1.2.

type Part = (Instrument, Subpart)

A part is a subdivided group of instruments of a given type.

TODO interaction between similar instruments, i.e. a trombone chorus of TTTB.

type Instrument = String

TODO

divide :: Int -> Part -> [Part]