Music.Pitch.Absolute

Description

Absolute pitch representation.

The canonical pitch representation is frequency in Hertz (Hz). For conversion, see HasFrequency.

Documentation

newtype Hertz

Absolute frequency in Hertz.

Constructors

Hertz

Fields

getHertz :: Double

data Cents

Number of pure octaves.

Cents are a logarithmic representation of frequency such that

 f * (2/1) = frequency (cents f + 1200)    

class HasFrequency a where

Methods

frequency :: a -> Hertz

cents :: HasFrequency a => a -> Cents