Music.Pitch.Common.Interval

Description

Provides standard intervals.

Intervals

data Interval

An interval is the difference between two pitches, incuding negative intervals.

Intervals and pitches can be added using .+^. To get the interval between two pitches, use .-..

 c .+^ minor third == eb
 f .-. c           == perfect fourth

Adding intervals preserves spelling. For example:

 m3 ^+^ _M3 = _P5
 d5 ^+^ _M6 = m10

The scalar type of intervals are Integer, using ^* to stack intervals of a certain type on top of each other. For example _P5 ^* 2 is a stack of 2 perfect fifths. The Num instance works as expected for +, negate and abs, and arbitrarily uses minor seconds for multiplication. If you find yourself *, or signum on intervals, consider switching to *^ or normalized.

Intervals are generally described in terms of Quality and Number. To construct an interval, use the interval constructor, the utility constructors major, minor, augmented and diminished, or the interval literals:

 m5  == minor   fifth    == interval Minor   5
 _P4 == perfect fourth   == interval Perfect 5
 d5  == diminished fifth == diminish (perfect fifth)

Creating intervals

interval :: Quality -> Number -> Interval

Creates an interval from a quality and number.

If given Perfect with an imperfect number (such as 3 or 7) a major interval is returned. If given Major or Minor with a perfect number (such as 5), constructs a perfect or diminished interval respectively.

perfect :: Number -> Interval

Creates a perfect interval. If given an inperfect number, constructs a major interval.

major :: Number -> Interval

Creates a major interval. If given a perfect number, constructs a perfect interval.

minor :: Number -> Interval

Creates a minor interval. If given a perfect number, constructs a diminished interval.

augmented :: Number -> Interval

Creates an augmented interval.

diminished :: Number -> Interval

Creates a diminished interval.

doublyAugmented :: Number -> Interval

Creates a doubly augmented interval.

doublyDiminished :: Number -> Interval

Creates a doubly diminished interval.

Inspecting intervals

isPositive :: Interval -> Bool

Returns whether the given interval is positive.

isNegative :: Interval -> Bool

Returns whether the given interval is negative.

Simple and compound intervals

isSimple :: Interval -> Bool

Returns whether the given interval is simple.

A simple interval is an positive interval spanning less than one octave.

isCompound :: Interval -> Bool

Returns whether the given interval is compound.

A compound interval is either a negative interval, or a positive interval spanning more than octave.

separate :: Interval -> (Octaves, Interval)

Separate a compound interval into octaves and a simple interval.

 (perfect octave)^*x + y = z  iff  (x, y) = separate z

simple :: Interval -> Interval

Returns the simple part of an interval.

 (perfect octave)^*x + y = z  iff  y = simple z

Inversion

invert :: Interval -> Interval

Intervallic inversion.

The inversion an interval is determined as follows:

The number of a simple interval the difference of nine and the number of its inversion.

The quality of a simple interval is the inversion of the quality of its inversion.

The inversion of a compound interval is the inversion of its simple component.

Utility

asInterval :: Interval -> Interval

This is just the identity function, but is useful to fix the type of Interval.

intervalDiff :: Interval -> Int

Deprecated: This should be hidden

interval' :: Int -> Int -> Interval

Deprecated: This should be hidden