Music.Pitch.Common.Number

Description

Provides interval numbers.

Number

data Number

The number portion of an interval (i.e. second, third, etc).

Note that the interval number is always one step larger than number of steps spanned by the interval (i.e. a third spans two diatonic steps). Thus number does not distribute over addition:

 number (a + b) = number a + number b - 1

class HasNumber a where

Methods

number :: a -> Number

Returns the number portion of an interval.

The interval number is negative if and only if the interval is negative.

See also quality, octaves and semitones.

unison :: Number

prime :: Number

second :: Number

third :: Number

fourth :: Number

fifth :: Number

sixth :: Number

seventh :: Number

octave :: Number

ninth :: Number

tenth :: Number

twelfth :: Number

duodecim :: Number

thirteenth :: Number

fourteenth :: Number

fifteenth :: Number

Inspecting

isConsonance :: HasNumber a => a -> Bool

isPerfectConsonance :: HasNumber a => a -> Bool

isImperfectConsonance :: HasNumber a => a -> Bool

isDissonance :: HasNumber a => a -> Bool