Music.Pitch.Common.Spell

Description

Provides standard pith names.

Spelling

type Spelling = Semitones -> Number

spell :: (Num a, HasSemitones p, p ~ Diff a, AffineSpace a) => Spelling -> a -> Pitch

spellInterval :: HasSemitones a => Spelling -> a -> Interval

modal :: Spelling

Spell pitches as c cs d eb e f fs g gs a bb b

sharps :: Spelling

Spell pitches as c cs d ds e f fs g gs a as b

flats :: Spelling

Spell pitches as c db d eb e f gb g ab a bb b