Music.Score.Annotations

Description

Provides annotations

Documentation

data Annotation

An annotation is a unique textual value attached to parts of a score. They are ignored by default, but can be collected with withAnnotations.

annotate :: String -> Score a -> Score a

Annotate the whole score.

annotateSpan :: Span -> String -> Score a -> Score a

Annotate a part of the score.

showAnnotations :: (HasPart' a, HasText a) => Score a -> Score a

Show all annotations in the score.

showAnnotations' :: (HasPart' a, HasText a) => String -> Score a -> Score a

Show all annotations in the score using the given prefix.

withAnnotations :: (HasPart' a, HasText a) => ([String] -> Score a -> Score a) -> Score a -> Score a

Handle the annotations in a score.