Music.Score.Chord

Description

Provides a representation for chords.

Chord representation

class HasChord a where

Associated Types

type Note a :: *

Methods

getChord :: a -> [Note a]

newtype ChordT a

Constructors

ChordT

Fields

getChordT :: [a]

Chord transformations

simultaneous :: Semigroup a => Score a -> Score a

Merge simultaneous events.

Two events are considered simultaneous iff their onset and offset are the same.

simultaneous' :: Score a -> Score [a]

Group simultaneous events.

Two events are considered simultaneous iff their onset and offset are the same.