Music.Score.Combinators

Description

Combinators for manipulating scores.

Rests

rest :: MonadPlus s => s (Maybe a)

Create a score containing a rest at time zero of duration one.

 Score (Maybe a)

noteRest :: MonadPlus s => Maybe a -> s a

Create a note or a rest. This is an alias for mfromMaybe with a nicer reading.

This function uses the unit position (0, 1).

 Maybe a -> Score a

removeRests :: MonadPlus s => s (Maybe a) -> s a

Remove all rests from a score. This is an alias for mcatMaybes with a nicer reading.

 Score (Maybe a) -> Score a

type Slicable a = (Ord (Duration a), AdditiveGroup (Duration a), Performable a, Composable a)

before :: Slicable a => Time a -> a -> a

Return a score containing only the notes whose offset falls before the given duration.

 Time -> Score a -> Score a

after :: Slicable a => Time a -> a -> a

Return a score containing only the notes whose onset falls after given duration.

 Time -> Score a -> Score a

slice :: Slicable a => Time a -> Time a -> a -> a

Return a score containing only the notes whose onset and offset falls between the given durations.

 Time -> Time -> Score a -> Score a

Maps and filters

Events

type Mappable a b = (Performable a, Composable b, Duration a ~ Duration b)

filterEvents :: (Performable a, Composable a) => (Time a -> Duration a -> Event a -> Bool) -> a -> a

Filter the events in a score.

 (Time -> Duration -> a -> Bool) -> Score a -> Score b

mapEvents :: Mappable a b => (Time a -> Duration a -> Event a -> Event b) -> a -> b

Map over the events in a score.

 (Time -> Duration -> a -> b) -> Score a -> Score b

mapFilterEvents :: Mappable a b => (Time a -> Duration a -> Event a -> Maybe (Event b)) -> a -> b

Efficient combination of mapEvents and filterEvents.

 (Time -> Duration -> a -> Maybe b) -> Score a -> Score b

Phrases

type Phraseable a b = (Mappable a b, Composable a, HasPart' (Event a))

mapFirst :: (Phraseable a b, e ~ Event a, e' ~ Event b) => (e -> e') -> (e -> e') -> a -> b

Map over the first, and remaining notes in each part.

If a part has only one notes, the first function is applied. If a part has no notes, the given score is returned unchanged.

 (a -> b) -> (a -> b) -> Score a -> Score b

mapLast :: (Phraseable a b, e ~ Event a, e' ~ Event b) => (e -> e') -> (e -> e') -> a -> b

Map over the last, and preceding notes in each part.

If a part has only one notes, the first function is applied. If a part has no notes, the given score is returned unchanged.

 (a -> b) -> (a -> b) -> Score a -> Score b

mapPhrase :: (Phraseable a b, e ~ Event a, e' ~ Event b) => (e -> e') -> (e -> e') -> (e -> e') -> a -> b

Map over the first, middle and last note in each part.

If a part has fewer than three notes the first takes precedence over the last, and last takes precedence over the middle.

 (a -> b) -> (a -> b) -> (a -> b) -> Score a -> Score b

mapPhraseSingle :: (Mappable a b, e ~ Event a, e' ~ Event b) => (e -> e') -> (e -> e') -> (e -> e') -> a -> b

Equivalent to mapPhrase for single-part scores.

Fails if the score contains overlapping events.

 (a -> b) -> (a -> b) -> (a -> b) -> Score a -> Score b

Parts

Extracting

filterPart :: (Mappable a a, Event a ~ e, HasPart' e) => (Part e -> Bool) -> a -> a

Filter a score to include only those events whose parts match a given predicate.

 (Part -> Bool) -> Score a -> Score a

extractParts :: (HasPart' e, Mappable a a, e ~ Event a) => a -> [a]

Extract parts from the a score.

The parts are returned in the order defined the associated Ord instance part type. You can recompose the score with mconcat, i.e.

 mconcat . extractParts = id

 Score a -> [Score a]

extractParts' :: (HasPart' e, Mappable a a, e ~ Event a) => a -> [(Part e, a)]

Extract parts from the a score and include the part name.

The parts are returned in the order defined the associated Ord instance part type.

Simple type

 Score a -> [(Part a, Score a)]

Mapping

mapPart :: (Enum a, Mappable b b, HasPart' e, e ~ Event b) => a -> (b -> b) -> b -> b

Map over a specific part in the given score.

 Part -> (Score a -> Score a) -> Score a -> Score a

mapParts :: (Monoid' b, Mappable a a, HasPart' e, e ~ Event a) => (a -> b) -> a -> b

Map over all parts in the given score.

 (Score a -> Score a) -> Score a -> Score a

mapAllParts :: (Monoid' b, Mappable a a, HasPart' e, e ~ Event a) => ([a] -> [b]) -> a -> b

Map over all parts in the given score.

 ([Score a] -> [Score a]) -> Score a -> Score a

Inspecting

getParts :: (Performable a, HasPart' e, e ~ Event a) => a -> [Part e]

Get all parts in the given score. Returns a list of parts.

 Score a -> [Part]

setParts :: (HasPart a, Functor s) => Part a -> s a -> s a

Set all parts in the given score.

 Part -> Score a -> Score a

modifyParts :: (Mappable a a, HasPart e, e ~ Event a) => (Part e -> Part e) -> a -> a

Modify all parts in the given score.

 (Part -> Part) -> Score a -> Score a

Part composition

(</>) :: (Enum (Part e), Mappable a a, HasPart' e, e ~ Event a) => a -> a -> a

Similar to <>, but increases parts in the second part to prevent collision.

moveParts :: (Enum (Part e), Integral b, Mappable a a, HasPart e, e ~ Event a) => b -> a -> a

Move down one voice (all parts).

moveToPart :: (Enum (Part e), Enum b, Mappable a a, HasPart e, e ~ Event a) => b -> a -> a

Move top-part to the specific voice (other parts follow).

Zippers

apply :: HasPart' a => Voice (Score a -> Score b) -> Score a -> Score b

Apply a time-varying function to all events in score.

snapshot :: HasPart' a => Score b -> Score a -> Score (b, Score a)

Get all notes that start during a given note.

applySingle :: Voice (Score a -> Score b) -> Score a -> Score b

Apply a time-varying function to all events in score.

snapshotSingle :: Score a -> Score b -> Score (a, Score b)

Get all notes that start during a given note.

Conversion

voiceToScore :: Voice a -> Score a

Convert a voice into a score.

voiceToScore' :: Voice (Maybe a) -> Score a

Convert a voice which may contain rests into a score.

scoreToVoice :: Score a -> Voice (Maybe a)

Convert a score into a voice.

This function fails if the score contain overlapping events.