Music.Score.Convert

Description

Provides conversion betwen Score, Voice and Track.

Conversion

voiceToScore :: Voice a -> Score a

Convert a voice into a score.

voiceToScore' :: Voice (Maybe a) -> Score a

Convert a voice which may contain rests into a score.

scoreToVoice :: Score a -> Voice (Maybe a)

Convert a score into a voice.

This function fails if the score contain overlapping events.