Music.Score.Export.Lilypond

Description

Documentation

type Lilypond = Music

class Tiable a => HasLilypond a where

Class of types that can be converted to Lilypond.

Methods

getLilypond :: Duration -> a -> Lilypond

Convert a value to a Lilypond music expression.

getLilypondChord :: Duration -> [a] -> Lilypond

getLilypondWithPrefix :: Duration -> a -> (Lilypond -> Lilypond, Lilypond)

getLilypondChordWithPrefix :: Duration -> [a] -> (Lilypond -> Lilypond, Lilypond)

toLilypond :: (HasLilypond a, HasPart' a, Semigroup a) => Score a -> Lilypond

Convert a score to a Lilypond representation.

toLilypondString :: (HasLilypond a, HasPart' a, Semigroup a) => Score a -> String

Convert a score to a Lilypond string.

showLilypond :: (HasLilypond a, HasPart' a, Semigroup a) => Score a -> IO ()

Convert a score to a Lilypond representaiton and print it on the standard output.

openLilypond :: (HasLilypond a, HasPart' a, Semigroup a) => Score a -> IO ()

Typeset a score using Lilypond and open it.

writeLilypond :: (HasLilypond a, HasPart' a, Semigroup a) => FilePath -> Score a -> IO ()

Convert a score to a Lilypond representation and write to a file.

Options

data LilypondOptions

Constructors

Inline

Score

writeLilypond' :: (HasLilypond a, HasPart' a, Semigroup a) => LilypondOptions -> FilePath -> Score a -> IO ()

Convert a score to a Lilypond representation and write to a file.

openLilypond' :: (HasLilypond a, HasPart' a, Semigroup a) => LilypondOptions -> Score a -> IO ()