Music.Score.Export.Midi

Description

Documentation

class HasMidi a where

Class of types that can be converted to MIDI.

Numeric types are interpreted as notes with a default velocity, pairs are interpreted as (pitch, velocity) pairs.

Minimal definition: getMidi. Given getMidiScore, getMidi can be implemented as getMidiScore . return.

Methods

getMidi :: a -> Score Message

Convert a value to a MIDI score. Typically, generates an on event using note followed by an optional off event.

getMidiScore :: Score a -> Score Message

Convert a score to a MIDI score. The default definition can be overriden for efficiency.

toMidi :: HasMidi a => Score a -> Midi

Convert a score to a MIDI file representation.

toMidiTrack :: HasMidi a => Score a -> Track Message

Convert a score to a track of MIDI messages.

writeMidi :: HasMidi a => FilePath -> Score a -> IO ()

Convert a score MIDI and write to a file.

playMidi :: HasMidi a => String -> Score a -> Event MidiMessage

Convert a score to a MIDI event.

playMidiIO :: HasMidi a => String -> Score a -> IO ()

Convert a score to a MIDI event and run it.