Music.Score.Export.MusicXml

Description

Documentation

type XmlScore = Score

type XmlMusic = Music

class Tiable a => HasMusicXml a where

Class of types that can be converted to MusicXML.

Methods

getMusicXml :: DurationT -> a -> XmlMusic

Convert a value to MusicXML.

Typically, generates a XmlMusic value using note or chord, and transforms it to add beams, slurs, dynamics, articulation etc.

getMusicXmlChord :: DurationT -> [a] -> XmlMusic

toXml :: (HasMusicXml a, HasPart' a, Show (Part a), Semigroup a) => Score a -> XmlScore

Convert a score to a MusicXML representation.

toXmlString :: (HasMusicXml a, HasPart' a, Show (Part a), Semigroup a) => Score a -> String

writeXml :: (HasMusicXml a, HasPart' a, Show (Part a), Semigroup a) => FilePath -> Score a -> IO ()

Convert a score to MusicXML and write to a file.

openXml :: (HasMusicXml a, HasPart' a, Show (Part a), Semigroup a) => Score a -> IO ()

Convert a score to MusicXML and open it.