Music.Score.Export.Util

Description

Documentation

voiceToBars :: Tiable a => Voice (Maybe a) -> [[(DurationT, Maybe a)]]

Convert a single-voice score to a list of bars.

separateBars :: Voice (Maybe a) -> [[(DurationT, Maybe a)]]

Given a set of absolute-time occurences, separate at each zero-time occurence. Note that this require every bar to start with a zero-time occurence.

toRelative :: [(TimeT, DurationT, b)] -> [(TimeT, DurationT, b)]

Convert absolute to relative durations.

spellPitch :: Integral a => a -> (a, a, a)

splitWhile :: (a -> Bool) -> [a] -> [[a]]

Group a list into sublists whereever a predicate holds. The matched element is the first in the sublist.

 splitWhile isSpace "foo bar baz"
    ===> ["foo"," bar"," baz"]

 splitWhile (> 3) [1,5,4,7,0,1,2]
    ===> [[1],[5],[4],[7,0,1,2]]

execute :: FilePath -> [String] -> IO ()

unRatio :: Integral t => Ratio t -> (t, t)

first :: (t -> t1) -> (t, t2) -> (t1, t2)

second :: (t -> t2) -> (t1, t) -> (t1, t2)

both :: (t -> t1) -> (t3 -> t2) -> (t, t3) -> (t1, t2)