Music.Score.Import.Sibelius

Documentation

type IsSibelius a = (IsPitch a, HasPart' a, Enum (Part a), HasPitch a, Num (PitchOf a), HasTremolo a, HasArticulation a, Tiable a)

This constraint includes all note types that can be constructed from a Sibelius representation.

fromSib :: IsSibelius a => SibScore -> Score a

Convert a score from a Sibelius representation.

readSib :: IsSibelius a => FilePath -> IO (Score a)

Read a Sibelius score from a file. Fails if the file could not be read or if a parsing error occurs.

readSibMaybe :: IsSibelius a => FilePath -> IO (Maybe (Score a))

Read a Sibelius score from a file. Fails if the file could not be read, and returns Nothing if a parsing error occurs.

readSibEither :: IsSibelius a => FilePath -> IO (Either String (Score a))

Read a Sibelius score from a file. Fails if the file could not be read, and returns Left m if a parsing error occurs.