Music.Score.Meta

Description

Provides meta-information.

Each score supports an unlimited number of Reactive meta-values.

This is more or less based on Diagrams styles, which is in turn based on XMonad.

Attributes

type IsAttribute a = (Typeable a, Monoid' a)

data Attribute

An existential wrapper type to hold attributes.

wrapAttr :: IsAttribute a => a -> Attribute

Wrap up an attribute.

unwrapAttr :: IsAttribute a => Attribute -> Maybe a

Meta-values

data Meta

addMetaNote :: forall a b. (IsAttribute a, HasMeta b, HasPart' b) => Note a -> b -> b

addMetaNoteNP :: forall a b. (IsAttribute a, HasMeta b) => Note a -> b -> b

runMeta :: forall a b. (HasPart' a, IsAttribute b) => Maybe a -> Meta -> Reactive b

class HasMeta a where

Type class for things which have meta-information.

Methods

applyMeta :: Meta -> a -> a

Apply meta-information by combining it (on the left) with the existing meta-information.

Specific meta-values

data Clef

Constructors

GClef

CClef

FClef

setClef :: (HasMeta a, HasPart' a, HasOnset a, HasOffset a) => Clef -> a -> a

setClefDuring :: (HasMeta a, HasPart' a) => Span -> Clef -> a -> a

data TimeSignature

setTimeSignature :: (HasMeta a, HasPart' a, HasOnset a, HasOffset a) => TimeSignature -> a -> a

setTimeSignatureDuring :: (HasMeta a, HasPart' a) => Span -> TimeSignature -> a -> a

data Fifths

data KeySignature

key :: Fifths -> Bool -> KeySignature

setKeySignature :: (HasMeta a, HasPart' a, HasOnset a, HasOffset a) => KeySignature -> a -> a

setKeySignatureDuring :: (HasMeta a, HasPart' a) => Span -> KeySignature -> a -> a

data Tempo

setTempo :: (HasMeta a, HasPart' a, HasOnset a, HasOffset a) => Tempo -> a -> a

setTempoDuring :: (HasMeta a, HasPart' a) => Span -> Tempo -> a -> a

data Title

mkTitle :: String -> Title

denoteTitle :: Title -> Title

getTitle :: Title -> [String]

getTitleAt :: Title -> Int -> Maybe String

title :: (HasMeta a, HasPart' a, HasOnset a, HasOffset a) => Title -> a -> a

titleDuring :: (HasMeta a, HasPart' a) => Span -> Title -> a -> a

subtitle :: (HasMeta a, HasPart' a, HasOnset a, HasOffset a) => Title -> a -> a

subtitleDuring :: (HasMeta a, HasPart' a) => Span -> Title -> a -> a

data Attribution

getAttribution :: Attribution -> String -> Maybe String

attribution :: [(String, String)] -> Attribution

attribution1 :: String -> String -> Attribution

composer :: (HasMeta a, HasPart' a, HasOnset a, HasOffset a) => String -> a -> a

composerDuring :: (HasMeta a, HasPart' a) => Span -> String -> a -> a