Music.Score.Meta.Clef

Description

Clef type

data Clef

Constructors

GClef

CClef

FClef

Adding clefs to scores

clef :: (HasMeta a, HasPart' a, HasOnset a, HasOffset a) => Clef -> a -> a

Set clef of the given score.

clefDuring :: (HasMeta a, HasPart' a) => Span -> Clef -> a -> a

Set clef of the given part of a score.

Extracting clefs

withClef :: HasPart' a => Clef -> (Clef -> Score a -> Score a) -> Score a -> Score a

Extract the clef in from the given score, using the given default clef.

The given function is called once for each clef change, containing the fragment of the score to which the given clef change is to be applied. This is mostly used by notation backends to emit a clef mark at the beginning of each fragment.