Music.Score.Meta.Fermata

Description

Fermata type

data FermataType

Constructors

StandardFermata

LongFermata

VeryLongFermata

data Fermata

Represents a fermata.

TODO where is the fermata added if the score contains multiple notes. Always the last?

Adding fermatas to scores

fermata :: (HasMeta a, HasPart' a, HasOnset a, HasOffset a) => Fermata -> a -> a

Add a fermata over the whole score.

fermataDuring :: (HasMeta a, HasPart' a) => Span -> Fermata -> a -> a

Add a fermata to the given score.

Extracting fermatas

withFermata :: (Fermata -> Score a -> Score a) -> Score a -> Score a

Extract fermatas in from the given score, using the given default fermata.