Music.Score.Meta.Time

Description

Time signature type

data TimeSignature

time :: Integer -> Integer -> TimeSignature

Create a simple time signature.

compoundTime :: [Integer] -> Integer -> TimeSignature

Create a compound time signature.

getTime :: TimeSignature -> ([Integer], Integer)

Extract the components of a time signature. Semantic function.

Adding time signature to scores

timeSignature :: (HasMeta a, HasPart' a, HasOnset a, HasOffset a) => TimeSignature -> a -> a

Set the sime signature of the given score.

timeSignatureDuring :: (HasMeta a, HasPart' a) => Span -> TimeSignature -> a -> a

Set the sime signature of the given part of a score.

Extracting time signatures

getTimeSignature :: TimeSignature -> Score a -> Reactive TimeSignature

getTimeSignatureChanges :: TimeSignature -> Score a -> [(Time, TimeSignature)]

withTimeSignature :: TimeSignature -> (TimeSignature -> Score a -> Score a) -> Score a -> Score a

Extract the time signature in from the given score, using the given default time signature.

The given function is called once for each time signature change, containing the fragment of the score to which the given time signature change is to be applied. This is mostly used by notation backends to emit a time signature mark at the beginning of each fragment.

Utility

getBarDurations :: [(TimeSignature, Duration)] -> [Duration]

getBarTimeSignatures :: [(TimeSignature, Duration)] -> [TimeSignature]

standardTimeSignature :: Duration -> TimeSignature

Time signature typically used for the given duration.