Music.Score.Ornaments

Description

Provides functions for manipulating ornaments (and miscellaneous stuff to be given its own module soon...).

Tremolo

class HasTremolo a where

Methods

setTrem :: Int -> a -> a

newtype TremoloT a

Constructors

TremoloT

Fields

getTremoloT :: (Int, a)

tremolo :: (Functor s, HasTremolo b) => Int -> s b -> s b

Set the number of tremolo divisions for all notes in the score.

Text

class HasText a where

Methods

addText :: String -> a -> a

newtype TextT a

Constructors

TextT

Fields

getTextT :: ([String], a)

text :: (Phraseable a a, HasPart' e, HasText e, e ~ Event a) => String -> a -> a

Attach the given text to the first note in the score.

Harmonics

class HasHarmonic a where

Methods

setHarmonic :: Int -> a -> a

newtype HarmonicT a

Constructors

HarmonicT

Fields

getHarmonicT :: (Int, a)

harmonic :: (Functor s, HasHarmonic a) => Int -> s a -> s a

Make all notes natural harmonics on the given overtone (1 for octave, 2 for fifth etc). Sounding pitch is unaffected, but notated output is transposed automatically.

artificial :: (Functor s, HasHarmonic a) => s a -> s a

Make all notes natural harmonics on the given overtone (1 for octave, 2 for fifth etc). Sounding pitch is unaffected, but notated output is transposed automatically.

Slides and glissando

class HasSlide a where

Methods

setBeginGliss :: Bool -> a -> a

setBeginSlide :: Bool -> a -> a

setEndGliss :: Bool -> a -> a

setEndSlide :: Bool -> a -> a

newtype SlideT a

Constructors

SlideT

Fields

getSlideT :: (Bool, Bool, a, Bool, Bool)

slide :: (Phraseable a a, Composable a, HasPart' e, HasSlide e, e ~ Event a) => a -> a

Slide between the first and the last note.