Music.Score.Part

Description

Provides functions for manipulating parts.

Part representation

class HasPart a where

Class of types with an associated part.

The part type can be any type that is orddered.

Associated Types

type Part a :: *

Associated part type. Should implement Ord and Show.

Methods

getPart :: a -> Part a

Get the voice of the given note.

setPart :: Part a -> a -> a

Set the voice of the given note.

modifyPart :: (Part a -> Part a) -> a -> a

Modify the voice of the given note.

type HasPart' a = (Ord (Part a), HasPart a)

Like HasPart, but enforces the part to be ordered. This is usually required for part separation and traversal.

newtype PartT n a

Constructors

PartT

Fields

getPartT :: (n, a)

getParts :: (Performable a, HasPart' e, e ~ Event a) => a -> [Part e]

Get all parts in the given score. Returns a list of parts.

 Score a -> [Part]

setParts :: (HasPart a, Functor s) => Part a -> s a -> s a

Set all parts in the given score.

 Part -> Score a -> Score a