Music.Score.Pitch

Description

Provides functions for manipulating pitch.

Accessors

pitch' :: HasPitch' a => Lens' a (Pitch a)

A lens to the pitch in a note, score or other structure.

pitch :: HasPitch a b => Lens a b (Pitch a) (Pitch b)

A lens to the pitch in a note, score or other structure.

Transformations

Transformations

inv :: Transposable a => Pitch a -> a -> a

Invert pitches.

Transformations

up :: Transposable a => Interval a -> a -> a

Transpose up.

down :: Transposable a => Interval a -> a -> a

Transpose down.

fifthsUp :: Transposable a => Scalar (Interval a) -> a -> a

Add the given interval below.

fifthsDown :: Transposable a => Scalar (Interval a) -> a -> a

Add the given interval below.

octavesUp :: Transposable a => Scalar (Interval a) -> a -> a

Transpose up by the given number of octaves.

octavesDown :: Transposable a => Scalar (Interval a) -> a -> a

Transpose down by the given number of octaves.

Transformations

fifthsAbove :: (Semigroup a, Transposable a) => Scalar (Interval a) -> a -> a

Add the given interval below.

fifthsBelow :: (Semigroup a, Transposable a) => Scalar (Interval a) -> a -> a

Add the given interval below.

octavesAbove :: (Semigroup a, Transposable a) => Scalar (Interval a) -> a -> a

Add the given interval below.

octavesBelow :: (Semigroup a, Transposable a) => Scalar (Interval a) -> a -> a

Add the given interval below.

Pitch representation

type family Pitch a

type Interval a = Diff (Pitch a)

class HasGetPitch s where

Class of types with readable pitch.

Methods

__getPitch :: a ~ Pitch s => s -> a

class SetPitch (Pitch t) s ~ t => HasSetPitch s t where

Class of types with mutable pitch.

Either setPitch or mapPitch can be implemented. If both are implemented, the following laws should be satisfied:

 setPitch x = mapPitch (const x)
 mapPitch f x = setPitch p x where p = f (__getPitch x)

For types that are Functors, the following instance can be used

 type instance Pitch (T a) = Pitch a
 instance HasSetPitch a b => HasSetPitch (T a) (T b) where
     type SetPitch g (T a) = T (SetPitch g a)
     mapPitch = fmap . mapPitch

Associated Types

type SetPitch b s :: *

Methods

__setPitch :: Pitch t -> s -> t

__mapPitch :: (Pitch s -> Pitch t) -> s -> t

type HasPitch' a = HasPitch a a

type HasPitch s t = (HasGetPitch s, HasSetPitch s t)

type HasSetPitch' a = HasSetPitch a a

type Transposable a = (HasSetPitch' a, Transposable1 a)

type Transposable1 a = (Diff (Pitch a) ~ Interval a, AffineSpace (Pitch a), VectorSpace (Interval a), IsPitch (Pitch a), IsInterval (Interval a))