Music.Score.Ties

Description

Provides a representation for tied notes, and a way to split a single note into a pair of tied notes.

Tiable class

class Tiable a where

Class of types that can be tied.

Methods

beginTie :: a -> a

Modify a note to be the first note in a tied note pair.

endTie :: a -> a

Modify a note to be the second note in a tied note pair.

toTied :: a -> (a, a)

Split a single note into a pair of tied notes.

The first returned element should have the original onset and the second element should have the original offset. Formally

 (onset . fst . toTied) a = onset a
 (offset . snd . toTied) a = offset a

newtype TieT a

Constructors

TieT

Fields

getTieT :: (Bool, a, Bool)

Splitting tied notes in scores

splitTies :: (HasPart' a, Tiable a) => Score a -> Score a

Split all notes that cross a barlines into a pair of tied notes.

splitTiesSingle :: Tiable a => Score a -> Score a

Equivalent to splitTies for single-voice scores. Fails if the score contains overlapping events.

splitTiesVoice :: Tiable a => Voice a -> Voice a

Split all notes that cross a barlines into a pair of tied notes.