Music.Time.Behavior

Description

Behaviors, or time-varying values.

TODO integrate better in the library

Documentation

data Behavior a

constant :: a -> Behavior a

A constant (non-varying) behavior.

Identical to behavior . const but creates more efficient behaviors.

behavior :: (Time -> a) -> Behavior a

Create a behavior from function of (absolute) time.

You should pass a function defined for the whole range, including negative time.

varying :: (Duration -> a) -> Behavior a

Create a behaviour from a function of (relative) duration focused on sunit.

You should pass a function defined for the whole range, including negative durations.

varyingIn :: Span -> (Duration -> a) -> Behavior a

Create a behaviour from a function of (relative) duration focused on the given span.

You should pass a function defined for the whole range, including negative durations.

switchB :: Time -> Behavior a -> Behavior a -> Behavior a

trimBeforeB :: Monoid a => Time -> Behavior a -> Behavior a