Music.Time.Onset

Description

This misleadingly named module provide a way to query a value for its duration, onset and offset.

Duration class

class HasDuration a where

Class of types with a duration.

If a type has an instance for both HasOnset and HasDuration, the following laws should hold:

 duration a = offset a .-- onset a

Methods

duration :: a -> Duration a

stretchTo :: (Stretchable a, HasDuration a, Fractional d, d ~ Duration a) => d -> a -> a

Stretch a score to fit into the given duration.

 Duration -> Score a -> Score a

Onset and offset class

class HasOnset a where

Class of types with a position in time.

Onset and offset are logical start and stop time, i.e. the preferred beginning and end of the sound, not o the the time of the attack and damp actions on an instrument,

If a type has an instance for both HasOnset and HasDuration, the following laws should hold:

 duration a = offset a .-- onset a

Methods

onset :: a -> Time a

Get the onset of the given value.

class HasOffset a where

Methods

offset :: a -> Time a

Get the offset of the given value.

startAt :: (HasOnset a, Delayable a, AffineSpace t, t ~ Time a) => t -> a -> a

Move a score so that its onset is at the specific time.

 Time -> Score a -> Score a

stopAt :: (HasOffset a, Delayable a, AffineSpace t, t ~ Time a) => t -> a -> a

Move a score so that its offset is at the specific time.

 Time -> Score a -> Score a

Utility

Default implementations

durationDefault :: (AdditiveGroup (Duration a), HasOffset a, HasOnset a) => a -> Duration a

Given HasOnset and HasOffset instances, this function implements duration.

onsetDefault :: (AdditiveGroup (Duration a), HasOffset a, HasDuration a) => a -> Time a

Given HasDuration and HasOffset instances, this function implements onset.

offsetDefault :: (AdditiveGroup (Duration a), HasOnset a, HasDuration a) => a -> Time a

Given HasOnset and HasOnset instances, this function implements offset.

Wrappers

newtype AddOffset t a

Constructors

AddOffset (t, a)