Music.Time.Performable

Description

Provides performable values. To perform a score is to render it as a list of time-duration-value triplets.

The compose function captures the fact that a value that is a Monoid, Transformable and Pointed can be reconstructed from a performance, i.e. compose is the inverse of perform.

Event and container types

type family Event s :: *

This type function returns the value type for a given type.

type family Container s :: * -> *

This type function returns the container type for a given type.

HasPoint class

The HasPoint class serves only to lift the Pointed class into unary kinds. We could have used Applicative or Monad here, but as we only care about pure and return, Pointed makes more sense.

Recall that:

 pure = return = point

type HasPoint a = (a ~ Container a (Event a), Pointed (Container a))

point' :: HasPoint a => Event a -> a

Equivalent to point, pure and return.

Performable values

class Performable a where

Performable values.

Minimal complete definition: perform.

Methods

perform :: a -> [(Time a, Duration a, Event a)]

Convert a score to a list of timed values.

The returned list must be sorted by comparing time.

performValues :: Performable a => a -> [Event a]

Return just the values of the score.

Composable values

class (Monoid' a, Transformable a, HasPoint a) => Composable a where

Composable values.

As one might expect

 compose . perform = id
 perform . compose = id

That is compose is the inverse of perform.

The methods in this class have a default implementation in terms of its superclasses, but are provided so that implementors may provide more efficient implementations.

Methods

event :: Time a -> Duration a -> Event a -> a

Creates a score with the given position and duration.

The default definition can be overridden for efficiency.

compose :: [(Time a, Duration a, Event a)] -> a

Creates a score from a list of timed values.

The given list need not be sorted.

The default definition can be overridden for efficiency.

Default implementations

memptyDefault :: Composable a => a

This function may be used as a value for mempty in a Monoid instance.

mappendDefault :: (Composable a, Performable a) => a -> a -> a

This function may be used as a value for mappend in a Monoid instance.

foldMapDefault :: (Monoid c, Performable a) => (Event a -> c) -> a -> c

This function may be used as a value for foldMap in a Foldable instance.