Music.Time.Reactive

Description

Reactive values, or piecewise functions of time.

Similar to Conal's definition in http://conal.net/blog/posts/reactive-normal-form, but defined in negative time as well. Its semantics function is either occs &&& ? or initial &&& updates, where intial is the value from negative infinity to the first update.

TODO integrate better in the library

Documentation

data Reactive a

Create, isos etc (TODO)

initial :: Reactive a -> a

Get the initial value.

final :: Reactive a -> a

Get the final value.

type Future a = (Time, a)

type Past a = (Time, a)

updates :: Reactive a -> [Future a]

Get the time of all updates and the value switched to at this point.

resets :: Reactive a -> [Past a]

occs :: Reactive a -> [Time]

Get the time of all updatese.

Predicates

isConstant :: Reactive a -> Bool

isVariable :: Reactive a -> Bool

Combinators

step :: (AdditiveGroup a, Fractional a) => Reactive a

The unit step function, which goes from 0 to 1 at start.

switch :: Time -> Reactive a -> Reactive a -> Reactive a

switch t a b behaves as a before time t, then as b.

trim :: Monoid a => Span -> Reactive a -> Reactive a

Replace everthing outside the given span by mempty.

trimBefore :: Monoid a => Time -> Reactive a -> Reactive a

Replace everthing before the given time by mempty.

trimAfter :: Monoid a => Time -> Reactive a -> Reactive a

Replace everthing after the given time by mempty.

Semantics