Music.Time.Reverse

Description

Provides reversible values.

Reversible class

class Reversible a where

Reversible values.

For instances of Reversible and HasOnset, the following laws should hold:

 onset a    = onset (rev a)
 duration a = duration (rev a)

For structural types, rev is applied recursively, hence the constraint on the Score instance. rev is id by default, so for a trivial type T it suffices to write

 instance Reversible T

For instances U of HasOnset and Transformable, a suitable instance is

 instance Reversible T where
     rev = withSameOnset (stretch (-1))

Methods

rev :: a -> a

Reverse a value.

Reverse is an involution, meaning that:

 rev (rev a) = a

Utility

newtype NoRev a

Constructors

NoRev

Fields

getNoRev :: a

newtype WithRev a

Constructors

WithRev (a, a)

withRev :: Reversible a => a -> WithRev a

fromWithRev :: Reversible a => WithRev a -> a