Music.Time.Span

Description

Provides time spans.

Documentation

data Span

A Span represents two points in time u and v where t <= u or, equivalently, a time t and a duration d where d >= 0.

A third way of looking at Span is that it represents a time transformation where onset is translation and duration is scaling.

Constructing spans

(<->) :: Time -> Time -> Span

t <-> u represents the span between t and u.

(-->) :: Time -> Duration -> Span

t --> d represents the span between t and t .+^ d.

Span as transformation

sunit :: Span

The default span, i.e. start <-> stop.

sapp :: (Delayable a, Stretchable a) => Span -> a -> a

Apply a span transformation.

sunder :: (Delayable a, Stretchable a) => Span -> (a -> a) -> a -> a

Apply a function under a span transformation.

sinvert :: Span -> Span

The inversion of a span.

 sinvert (sinvert s) = s
 sapp (sinvert s) . sapp s = id

Deconstructing spans

range :: Span -> (Time, Time)

Render a span as onset and offset.

delta :: Span -> (Time, Duration)

Render a span as a time and duration.

mapRange :: (Time -> Time -> (Time, Time)) -> Span -> Span

Map over the span as a time pair.

mapDelta :: (Time -> Duration -> (Time, Duration)) -> Span -> Span

Map over the span as onset and duration.