Music.Time.Time

Description

Time and duration types

type Time = Point Duration

This type represents absolute time in seconds since the start time. Note that time can be negative, representing events occuring before the start time. The start time is usually the the beginning of the musical performance.

Time forms an affine space with durations as the underlying vector space, that is, we can add a time to a duration to get a new time using .+^, take the difference of two times to get a duration using .-..

newtype Duration

This type represents relative time in seconds.

Constructors

Duration

Fields

getDuration :: Rational

Note that you should use .-. and .+^ to convert between time and duration. To refer to time zero (the beginning of the music), use origin.