Music.Parts

Description

Part representation.

Terminology

Parts represent a subset of a group of performers. It is mainly used for instrumental and vocal music, but some concetps may be useful in electronic music as well.

Section refers to a set of instrumentfamilies related by sound production method (i.e. woodwind).

Family refers to a set of instrument or voice types, which typically differ in size (i.e. saxophones).

Instrument refers to a set of instruments or voice types of a given type (i.e. soprano saxophones). Perhaps confusingly, this includes vocal types such as alto, tenor etc as well. However, there is no good general term that incorporate both instrument and voice type.

A Part is made up of an Instrument and a Division (i.e. Violin I). Solo parts are treated separately, so i.e. Violin solo II (as in a double concerto) is distinct from Violin II.

Subparts

data Division

A division represents a subset of a finite group of performers.

For example a group may be divided into three equal divisions, designated (0, 3), (1, 3) and (2, 3) respectively.

divisions :: Int -> [Division]

Get all possible divisions for a given divisor in ascending order.

getDivision :: Division -> (Int, Int)

data Subpart

A subpart is a potentially infinite sequence of divisions, each typically designated using a new index type, i.e. I.1.2.

The empty subpart (also known as def) represents all the players of the group, or in the context of Part, all players of the given instrument.

Instruments

data Instrument

An Instrument represents the set of all instruments of a given type.

Constructors

StdInstrument Int

OtherInstrument String

Parts

data Solo

Constructors

Solo

Tutti

data Part

A part is a subdivided group of instruments of a given type.

Constructors

Part Solo Instrument Subpart

divide :: Int -> Part -> [Part]

Divide a part inton subparts.

containsPart :: Part -> Part -> Bool

a `containsPart` b holds if the set of players represented by a is an improper subset of the set of players represented by b.

containsSubpart :: Subpart -> Subpart -> Bool

solo :: Instrument -> Part

tutti :: Instrument -> Part

Instruments etc

violin :: Instrument

viola :: Instrument

cello :: Instrument

bass :: Instrument

tubularBells :: Instrument

Default values

defaultClef :: Part -> Int

defaultMidiProgram :: Part -> Int

defaultMidiChannel :: Part -> Int

defaultMidiNote :: Part -> Int

Basic

data BasicPart