Music.Score.Combinators

Description

Combinators for manipulating scores and related structures.

Rests

note :: Monad s => a -> s a

Create a score containing a note at time zero and duration one.

This is an alias for return.

 a -> Score a

rest :: MonadPlus s => s (Maybe a)

Create a score containing a rest at time zero and duration one.

This is an alias for return Nothing.

 Score (Maybe a)

noteRest :: MonadPlus s => Maybe a -> s a

Create a note or a rest at time zero and duration one.

This is an alias for mfromMaybe with a nicer reading.

 Maybe a -> Score a

removeRests :: MonadPlus s => s (Maybe a) -> s a

Remove all rests from a score.

This is an alias for mcatMaybes with a nicer reading.

 Score (Maybe a) -> Score a

Maps and filters

Events

filterEvents :: (Time -> Duration -> a -> Bool) -> Score a -> Score a

Filter the events in a score.

mapEvents :: (Time -> Duration -> a -> b) -> Score a -> Score b

Map over the events in a score.

mapFilterEvents :: (Time -> Duration -> a -> Maybe b) -> Score a -> Score b

Efficient combination of mapEvents and filterEvents.

Editing

before :: Time -> Score a -> Score a

Return a score containing only the notes whose offset falls before the given duration.

after :: Time -> Score a -> Score a

Return a score containing only the notes whose onset falls after given duration.

slice :: Time -> Time -> Score a -> Score a

Return a score containing only the notes whose onset and offset falls between the given durations.

Map over phrases

mapFirst :: HasPart' a => (a -> b) -> (a -> b) -> Score a -> Score b

Map over the first, and remaining notes in each part.

If a part has only one notes, the first function is applied. If a part has no notes, the given score is returned unchanged.

mapLast :: HasPart' a => (a -> b) -> (a -> b) -> Score a -> Score b

Map over the last, and preceding notes in each part.

If a part has only one notes, the first function is applied. If a part has no notes, the given score is returned unchanged.

mapPhrase :: HasPart' a => (a -> b) -> (a -> b) -> (a -> b) -> Score a -> Score b

Map over the first, middle and last note in each part.

If a part has fewer than three notes the first takes precedence over the last, and last takes precedence over the middle.

mapPhraseSingle :: (a -> b) -> (a -> b) -> (a -> b) -> Score a -> Score b

Equivalent to mapPhrase for single-part scores.

Fails if the score contains overlapping events.

 (a -> b) -> (a -> b) -> (a -> b) -> Score a -> Score b

Parts

Extracting parts

filterPart :: HasPart' a => (Part a -> Bool) -> Score a -> Score a

Filter a score to include only those events whose parts match a given predicate.

extractParts :: HasPart' a => Score a -> [Score a]

Extract parts from the a score.

The parts are returned in the order defined the associated Ord instance part type. You can recompose the score with mconcat, i.e.

 mconcat . extractParts = id

extractParts' :: HasPart' a => Score a -> [(Part a, Score a)]

Extract parts from the a score and include the part name.

The parts are returned in the order defined the associated Ord instance part type.

Simple type

Map over parts

mapPart :: (Enum (Part a), HasPart' a) => Part a -> (Score a -> Score a) -> Score a -> Score a

Map over a specific part in the given score.

mapParts :: HasPart' a => (Score a -> Score b) -> Score a -> Score b

Map over all parts in the given score.

 (Score a -> Score a) -> Score a -> Score a

mapAllParts :: HasPart' a => ([Score a] -> [Score b]) -> Score a -> Score b

Map over all parts in the given score.

 ([Score a] -> [Score a]) -> Score a -> Score a

modifyParts :: HasPart' a => (Part a -> Part a) -> Score a -> Score a

Modify all parts in the given score.

 (Part -> Part) -> Score a -> Score a

Part composition

(</>) :: (Enum (Part a), Ord (Part a), HasPart a) => Score a -> Score a -> Score a

Similar to <>, but increases parts in the second part to prevent collision.

(/) :: (HasPart' a) => a -> a -> a

moveParts :: (Enum (Part a), Integral b, Ord (Part a), HasPart a) => b -> Score a -> Score a

Move down one voice (all parts).

moveParts :: (Enum (Part e), Integral b, Mappable a a, HasPart e, e ~ Event a) => b -> a -> a

moveToPart :: (Enum (Part a), Enum a1, Ord (Part a), HasPart a) => a1 -> Score a -> Score a

Move top-part to the specific voice (other parts follow).

moveToPart :: (Enum (Part e), Enum b, Mappable a a, HasPart e, e ~ Event a) => b -> a -> a

Zippers

apply :: HasPart' a => Voice (Score a -> Score b) -> Score a -> Score b

Apply a time-varying function to all events in score.

snapshot :: HasPart' a => Score b -> Score a -> Score (b, Score a)

Get all notes that start during a given note.

applySingle :: Voice (Score a -> Score b) -> Score a -> Score b

Apply a time-varying function to all events in score.

snapshotSingle :: Score a -> Score b -> Score (a, Score b)

Get all notes that start during a given note.