Music.Score.Export.Lilypond

Description

Documentation

type Lilypond = Music

class Tiable a => HasLilypond a where

Class of types that can be converted to Lilypond.

Methods

getLilypond :: DurationT -> a -> Lilypond

Convert a value to a Lilypond music expression.

getLilypondChord :: DurationT -> [a] -> Lilypond

toLy :: (HasLilypond a, HasPart' a, Show (Part a)) => Score a -> Lilypond

Convert a score to a Lilypond representation.

writeLy :: (HasLilypond a, HasPart' a, Show (Part a)) => FilePath -> Score a -> IO ()

Convert a score to a Lilypond representation and write to a file.

openLy :: (HasLilypond a, HasPart' a, Show (Part a)) => Score a -> IO ()

Typeset a score using Lilypond and open it.