Music.Score.Meta

Description

Provides meta-information.

Each score supports an unlimited number of Reactive meta-values.

This is more or less based on Diagrams styles, which is in turn based on XMonad.

Attributes

type IsAttribute a = (Typeable a, Monoid' a)

data Attribute

An existential wrapper type to hold attributes.

wrapAttr :: IsAttribute a => a -> Attribute

Wrap up an attribute.

unwrapAttr :: IsAttribute a => Attribute -> Maybe a

Meta-values

data Meta

addMetaNote :: forall a b. (IsAttribute a, HasMeta b, HasPart' b) => Note a -> b -> b

addGlobalMetaNote :: forall a b. (IsAttribute a, HasMeta b) => Note a -> b -> b

runMeta :: forall a b. (HasPart' a, IsAttribute b) => Maybe a -> Meta -> Reactive b

class HasMeta a where

Type class for things which have meta-information.

Methods

meta :: Lens' a Meta

Apply meta-information by combining it (on the left) with the existing meta-information.