Music.Time.Delayable

Description

Provides delayable values.

Delayable class

class Delayable a where

Delayable values.

Methods

delay :: Duration -> a -> a

Delay a value.

 Duration -> Score a -> Score a

move :: Delayable a => Duration -> a -> a

Move a score forward in time. Equivalent to 'delayTime.

 Duration -> Score a -> Score a

moveBack :: Delayable a => Duration -> a -> a

Move a score backward in time. Negated verison of 'delayTime

 Duration -> Score a -> Score a

Utility functions

delayTime :: Delayable a => Time -> a -> a

Delay relative to origin. Provided for situations when you have a value that should forward based on the distance between some time t and the origin, but it does not necessarily have a start time.

 Time -> Score a -> Score a