Music.Time.Reactive

Description

Reactive values.

Based on Conal's semantifcs as per http://conal.net/blog/posts/reactive-normal-form.

TODO integrate better in the library

Documentation

data Reactive a

occs :: Reactive a -> [Time]

(?) :: Reactive a -> Time -> a

switch :: Time -> Reactive a -> Reactive a -> Reactive a

switch t a b behaves as a before time t, then as b.

activate :: Note (Reactive a) -> Reactive a -> Reactive a

noteToReactive :: Monoid a => Note a -> Reactive a

initial :: Reactive a -> a

updates :: Reactive a -> [(Time, a)]

splitReactive :: Reactive a -> Either a ((a, Time), [Note a], (Time, a))

Split a reactive into notes, as well as the values before and after the first/last update