Music.Time.Time

Description

Time and duration types

type Time a = Point (Duration a)

This type function returns the duration type for a given type.

type family Duration s :: *

This type function returns the duration type for a given type.

Note that you should use .-. and .+^ to convert between time and duration. To refer to time zero (the beginning of the music), use origin.

Basic time and duration types

type DurationT = Rational

This type represents relative time in seconds.

type TimeT = Point DurationT

This type represents absolute time in seconds since the start time. Note that time can be negative, representing events occuring before the start time. The start time is usually the the beginning of the musical performance.

Time forms an affine space with durations as the underlying vector space, that is, we can add a time to a duration to get a new time using .+^, take the difference of two times to get a duration using .-..